test: fix test-without-async-context-frame.mjs in debug mode
The test is spawning the python test runner directly without
considering whether the current build is a debug build or not
i.e. does not propage the mode parameter when the test is
run with --mode=debug, so it always ends up using the release
build to run the test even if the test is being run for debug
builds. This patch fixes that.

PR-URL: https://github.com/nodejs/node/pull/57034
Reviewed-By: Richard Lau <rlau@redhat.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>